Better solution to build problem on windows

+++ src/plastimatch/base/rtplan_control_pt.h
@@ -22,12 +22,12 @@ public:
     std::string beam_limiting_device_rotation_direction;
 
     std::string scan_spot_tune_id;
-    unsigned long number_of_scan_spot_positions;
+    size_t number_of_scan_spot_positions;
     std::string scan_spot_reordering_allowed;
     std::vector<float> scan_spot_position_map;
     std::vector<float> scan_spot_meterset_weights;
 
-    unsigned long number_of_paintings;
+    size_t number_of_paintings;
     float scanning_spot_size[2];
     
     float patient_support_angle;
